business rules added
Added attribute for employee "emp_is_over_14" as a boolean value that tracks this. I would have implemented this as a CHECK statement as part of emp_birthdate, but couldn't reference the current date/time from the CHECK statement.

Added CHECK-s to product.price and order_header.total so that they can't be negative.

Added NOT NULL constraints to loyalty_member fields to comply with business rule.

Added query to demonstrate underage employee, and added sample underage employee that will show up in that query.

cafe_dump is refreshed to reflect these changes.